Interfaces injection added
