+++ services/product-service/product-impl/src/main/java/com/technokratos/eateasy/product/repository/CategoryRepository.java
@@ -1,68 +1,20 @@
 package com.technokratos.eateasy.product.repository;
 
 import com.technokratos.eateasy.product.entity.Category;
-import com.technokratos.eateasy.product.util.QueryProvider;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
 
-@Repository
-@RequiredArgsConstructor
-public class CategoryRepository {
+public interface CategoryRepository {
+  public void assignCategoriesToProduct(List<UUID> categoriesId, UUID productId);
 
-  private final JdbcTemplate jdbcTemplate;
-  private final QueryProvider queryProvider;
+  public List<Category> getCategoriesByProductId(UUID productId);
 
-  private final RowMapper<Category> categoryRowMapper =
-      (rs, rowNum) ->
-          Category.builder()
-              .id(UUID.fromString(rs.getString("id")))
-              .title(rs.getString("title"))
-              .build();
+  public Category save(Category category);
 
-  public void assignCategoriesToProduct(List<UUID> categoriesId, UUID productId) {
-    String sql = queryProvider.getSqlQueryForCategory("assign_categories_to_product");
-    List<Object[]> batchArgs =
-        categoriesId.stream().map(categoryId -> new Object[] {productId, categoryId}).toList();
-    jdbcTemplate.batchUpdate(sql, batchArgs);
-  }
+  public List<Category> findAll();
 
-  public List<Category> getCategoriesByProductId(UUID productId) {
-    String sql = queryProvider.getSqlQueryForCategory("get_categories_by_product_id");
-    return jdbcTemplate.query(sql, categoryRowMapper, productId);
-  }
+  public Optional<Category> findById(UUID id);
 
-  public Category save(Category category) {
-    String sql = queryProvider.getSqlQueryForCategory("save");
-    try {
-      return jdbcTemplate.queryForObject(sql, categoryRowMapper, category.getTitle());
-    } catch (DataIntegrityViolationException e) {
-      throw e;
-    } catch (Exception e) {
-      throw new RuntimeException("Database error", e);
-    }
-  }
-
-  public List<Category> findAll() {
-    String sql = queryProvider.getSqlQueryForCategory("find_all");
-    return jdbcTemplate.query(sql, categoryRowMapper);
-  }
-
-  public Optional<Category> findById(UUID id) {
-    return jdbcTemplate
-        .query(queryProvider.getSqlQueryForCategory("find_by_id"), categoryRowMapper, id)
-        .stream()
-        .findFirst();
-  }
-
-  public void updateCategoriesByProductId(List<UUID> categories, UUID productId) {
-    jdbcTemplate.update(
-        queryProvider.getSqlQueryForCategory("delete_from_summary_table"), productId);
-    assignCategoriesToProduct(categories, productId);
-  }
+  public void updateCategoriesByProductId(List<UUID> categories, UUID productId);
 }
